有福利关联类别时，不能删除类别

+++ src/main/java/com/manage/controller/WelfareCategoryController.java
@@ -3,9 +3,13 @@ package com.manage.controller;
 import cn.hutool.core.util.StrUtil;
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.baomidou.mybatisplus.core.toolkit.Wrappers;
+import com.manage.common.BusinessException;
+import com.manage.common.ResultCodeEnum;
+import com.manage.model.Welfare;
 import com.manage.model.WelfareCategory;
 import com.manage.model.comm.R;
 import com.manage.service.WelfareCategoryService;
+import com.manage.service.WelfareService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.*;
 
@@ -22,6 +26,9 @@ public class WelfareCategoryController {
 
     @Resource
     WelfareCategoryService welfareCategoryService;
+
+    @Resource
+    WelfareService welfareService;
 
     @GetMapping("/list")
     public R getAllCategories(@RequestParam(value = "categoryName", required = false) String categoryName) {
@@ -44,6 +51,10 @@ public class WelfareCategoryController {
 
     @DeleteMapping("/delete/{id}")
     public R delete(@PathVariable String id) {
+        Welfare welfare = welfareService.getOne(new LambdaQueryWrapper<Welfare>().eq(Welfare::getCategoryId, id));
+        if (welfare != null) {
+            throw new BusinessException(ResultCodeEnum.PARAM_ERROR, "已有福利关联该类别，取消关联后在删除");
+        }
         boolean res = welfareCategoryService.removeById(id);
         return res ? R.ok() : R.error();
